+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add: consider `cpsAcceleration` of multiple buying until tier upgrade (X buys to tier amount + buy tier upgrade + Y buys after tier upgrade)
 
 ## [2042.12] - 2021-09-21
 ### Added